+++ start/src/App.vue
@@ -1,12 +1,32 @@
 <script lang="ts" setup></script>
 
 <template>
-  <f-sticky-card :open="false" close-text="展开代码" open-text="折叠代码">
+  <f-list>
+    <template #header>
+      <div>111</div>
+      <div>111</div>
+      <div>111</div>
+    </template>
+
+    <template #footer>
+      <div>111</div>
+      <div>111</div>
+      <div>111</div>
+    </template>
+
+    <f-list-item>这是一段文字</f-list-item>
+    <f-list-item>这是一段文字</f-list-item>
+    <f-list-item>这是一段文字</f-list-item>
+    <f-list-item>这是一段文字</f-list-item>
+    <f-list-item>这是一段文字</f-list-item>
+  </f-list>
+
+  <!-- <f-sticky-card :open="false" close-text="展开代码" open-text="折叠代码">
     <template #source> click me </template>
 
     click me
     <p><em>content</em></p>
-  </f-sticky-card>
+  </f-sticky-card> -->
   <!-- <f-sticky-card :open="false" close-text="展开代码" open-text="折叠代码">
     <template #source>
       <f-button disabled type="default">默认按钮</f-button>